Moved cu_utils/linalg to the gef_cuda directory - Conditional import when using CUDA
